requirements.txt
python-dotenv
numpy
pandas
certifi
pymongo
pymongo[srv]==3.12
scikit-learn
dill
PyYAML
mlflow==2.8.1
setuptools>=65.0.0
dagshub
fastapi
uvicorn
# -e .